+++ src/main/java/com/company/ClinicaOdontologicaV1/service/impl/PacienteService.java
@@ -3,10 +3,12 @@ package com.company.ClinicaOdontologicaV1.service.impl;
 
 import com.company.ClinicaOdontologicaV1.dto.PacienteDTO;
 import com.company.ClinicaOdontologicaV1.entity.Paciente;
+import com.company.ClinicaOdontologicaV1.exceptions.ResourceNotFoundException;
 import com.company.ClinicaOdontologicaV1.repository.IPacienteRepository;
 import com.company.ClinicaOdontologicaV1.service.IPacienteService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,13 +16,13 @@ import org.springframework.stereotype.Service;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class PacienteService implements IPacienteService {
 
     private final IPacienteRepository pacienteRepository;
     private final ObjectMapper mapper;
+    private static final Logger LOGGER = Logger.getLogger(PacienteService.class);
 
     @Autowired
     public PacienteService(IPacienteRepository pacienteRepository) {
@@ -32,27 +34,41 @@ public class PacienteService implements IPacienteService {
     public void guardar(Paciente paciente) {
         paciente.setFechaIngreso(LocalDate.now());
         pacienteRepository.save(paciente);
+        LOGGER.info("Nuevo paciente registrado con exito.");
     }
 
     @Override
-    public PacienteDTO buscarPorId(Long id) throws Exception {
+    public PacienteDTO buscarPorId(Long id) throws ResourceNotFoundException {
         mapper.registerModule(new JavaTimeModule());
         Paciente encontrado =pacienteRepository.findById(id).orElse(null);
         if(encontrado!=null){
+            LOGGER.info("paciente encontrado: "+encontrado);
             return mapper.convertValue(encontrado,PacienteDTO.class);
         }else{
-            throw new Exception("Paciente no encontrado");
+            LOGGER.error("El id="+id+" del paciente no se encuentra en la base de datos.");
+            throw new ResourceNotFoundException("Paciente no encontrado");
         }
     }
 
     @Override
-    public void modificar(Paciente paciente) {
-        guardar(paciente);
+    public void modificar(Paciente paciente) throws ResourceNotFoundException {
+        if(buscarPorId(paciente.getId())!=null){
+            LOGGER.info("paciente "+paciente.getId()+" modificado con exito.");
+            guardar(paciente);
+        }else{
+            LOGGER.error("No se pudo modificar el paciente de id: "+paciente.getId());
+        }
     }
 
     @Override
-    public void eliminar(Long id) {
-        pacienteRepository.deleteById(id);
+    public void eliminar(Long id) throws ResourceNotFoundException {
+        if(buscarPorId(id)!=null){
+            pacienteRepository.deleteById(id);
+            LOGGER.warn("paciente id: "+id+" eliminado con exito.");
+        }else{
+            LOGGER.error("No se pudo eliminar el paciente de id: "+id);
+        }
+
     }
 
     @Override
@@ -63,6 +79,7 @@ public class PacienteService implements IPacienteService {
             System.out.println(p.getDomicilio());
             listaDTO.add(mapper.convertValue(p,PacienteDTO.class));
         }
+        LOGGER.info("Se solicito la lista de todos los pacientes.");
         return listaDTO;
     }
 }
